+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+[4.1.0](https://github.com/conductorphp/conductor-ssh-support/compare/4.0.0...4.1.0) (2026-08-10)
+
+### Features
+* PHP 8.4.1+ (CTAP-1224) ([20a7695](https://github.com/conductorphp/conductor-ssh-support/commit/20a76957f2ffabc2ea7ad1131a8f020cb9c616c8))
+* laminas-servicemanager v4 (CTAP-1221) ([c5bde24](https://github.com/conductorphp/conductor-ssh-support/commit/c5bde24b00f6422a0de845a3f35745a51ad85f15))
+
+<!--- CHANGELOG SPLIT MARKER -->
+
+[3.0.0](https://github.com/conductorphp/conductor-ssh-support/compare/2.0.0...3.0.0) (2026-06-26)
+
+### Bug Fixes
+* to phpseclib 3 and conductor 2.0 (CTAP-776) ([b665968](https://github.com/conductorphp/conductor-ssh-support/commit/b665968dbb5d3016379e7244b6fb766a67570f3d))
+
+<!--- CHANGELOG SPLIT MARKER -->
+
 [1.0.1](https://github.com/conductorphp/conductor-ssh-support/compare/1.0.0...1.0.1) (2026-06-25)
 
 ### Bug Fixes